Style: Gmail DTO 카테고리 별 디렉토리 생성 후 파일이동
