+++ src/main/java/woozlabs/echo/domain/gmail/dto/thread/GmailThreadGetPayload.java
@@ -1,8 +1,6 @@
-package woozlabs.echo.domain.gmail.dto;
+package woozlabs.echo.domain.gmail.dto.thread;
 
 import com.google.api.services.gmail.model.MessagePart;
-import com.google.api.services.gmail.model.MessagePartBody;
-import lombok.Builder;
 import lombok.Getter;
 
 import java.util.ArrayList;
