+++ src/main/java/woozlabs/echo/domain/gmail/dto/thread/GmailThreadListResponse.java
@@ -1,13 +1,10 @@
-package woozlabs.echo.domain.gmail.dto;
+package woozlabs.echo.domain.gmail.dto.thread;
 
 import lombok.Builder;
 import lombok.Getter;
 import woozlabs.echo.global.dto.ResponseDto;
 
-import java.text.SimpleDateFormat;
 import java.util.List;
-import java.util.Locale;
-import java.util.TimeZone;
 
 @Getter
 @Builder
